requirements.txt
# bpy==2.82.1
PyYAML==6.0
matplotlib